requirements.txt
boto3>=1.28.0
python-dotenv>=0.19.0
fastapi>=0.68.0
motor>=3.0.0
pydantic>=1.8.0